Restructure and add write declarations
