+++ docs/libx/bunny.d.ts
@@ -1,0 +1,3 @@
+declare export const positions: [number, number, number][];
+declare export const normals: [number, number, number][];
+declare export const elements: [number, number, number][];
