+++ docs/libx/cube.js
