+++ docs/libx/load-image.js
