+++ docs/libx/square.d.ts
@@ -1,0 +1,3 @@
+declare export const positions: [number, number, number][];
+declare export const uvs: [number, number][];
+declare export const elements: [number, number, number][];
