+++ docs/libx/square.js
@@ -5,14 +5,7 @@ export const positions = [
     [-1, -1],
 ]
 
-export const colors = [
-    [1, 0, 0, 1],
-    [0, 1, 0, 1],
-    [0, 0, 1, 1],
-    [1, 1, 0, 1],
-]
-
-export const texCoords = [
+export const uvs = [
     [1, 1],
     [0, 1],
     [1, 0],
