+++ docs/libx/teapot.d.ts
@@ -1,0 +1,4 @@
+declare export const positions: [number, number, number][];
+declare export const normals: [number, number, number][];
+declare export const elements: [number, number, number][];
+
