+++ docs/libx/uv-cube.d.ts
@@ -1,0 +1,4 @@
+declare export const positions: [number, number, number][];
+declare export const uvs: [number, number][];
+declare export const elements: [number, number, number][];
+
